pyproject.toml
[project]
name = "rl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ale-py>=0.10.2",
    "autorom[accept-rom-license]>=0.6.1",
    "flappy-bird-gymnasium>=0.4.0",
    "gym>=0.26.2",
    "gymnasium[atari]>=1.1.0",
    "numpy>=2.2.3",
    "pygame-ce>=2.5.3",
]
